style(qdf): compare object number in usize for the 1..N order check
Address Gemini review (PR #431, qdf_fix.rs:444): compare obj.num (u32)
against the index in usize rather than widening both sides to u64. usize
is the natural type for an index comparison and u32->usize is lossless on
flpdf's 32/64-bit targets; i + 1 cannot overflow as i is a Vec index. No
behavioral change.

+++ crates/flpdf/src/qdf_fix.rs
@@ -441,7 +441,7 @@ pub fn fix_qdf(input: &[u8]) -> Result<Vec<u8>> {
     // `/Length` holder inline after its stream (flpdf-abu3 / PR #430), so this
     // rejects nothing the writer (or qpdf `--qdf`) produces.
     for (i, obj) in objects.iter().enumerate() {
-        if u64::from(obj.num) != i as u64 + 1 {
+        if obj.num as usize != i + 1 {
             return Err(Error::parse(
                 obj.obj_line_start,
                 "fix_qdf: non-sequential object numbering \
